Renaming aux.js to util.js for Windows compatibility.  aux.* is forbidden in Windows environments.
