+++ lib/five.js
@@ -28,12 +28,12 @@
 // THIS SOFTWARE, EVEN IF ADVISED OF THE POSSIBILITY OF SUCH DAMAGE.
 
 var qt = require('node-qt'),
-    aux = require('./aux'),
+    util = require('./util'),
     five = require('./core');
 
 // Expose key and button codes, e.g. five.Key_Left, five.LeftButton, etc
-aux.extend(five, qt.Key);
-aux.extend(five, qt.MouseButton);
+util.extend(five, qt.Key);
+util.extend(five, qt.MouseButton);
 
 //
 // Expose classes
